requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.1
django-celery-beat==2.5.0
redis==5.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
